+++ include/point.h
@@ -12,7 +12,7 @@ class Point3D final
 {
 public:
     Point3D() = default;
-    Point3D(float x, float y, float z) : x(x), y(y), z(z) {}
+    Point3D(double x, double y, double z) : x(x), y(y), z(z) {}
 
     bool valid() const;
     bool equal(const Point3D& another_point) const;
@@ -23,9 +23,9 @@ public:
     Point3D operator-(const Point3D& point) const;
     void  operator=(const Vector3D& vector);
 
-    float x = NAN;
-    float y = NAN;
-    float z = NAN;
+    double x = NAN;
+    double y = NAN;
+    double z = NAN;
 };
 
 bool IsThreePointsOnOneLine(const Point3D& p0, const Point3D& p1, const Point3D& p2);
@@ -40,7 +40,7 @@ class Point2D final
 {
 public:
     Point2D() = default;
-    Point2D(float x, float y) : x(x), y(y) {}
+    Point2D(double x, double y) : x(x), y(y) {}
 
     bool valid() const;
     bool equal(const Point2D& another_point) const;
@@ -51,8 +51,8 @@ public:
     Point2D operator-(const Point2D& point) const;
     void    operator=(const Vector2D& vector);
 
-    float x = NAN;
-    float y = NAN;
+    double x = NAN;
+    double y = NAN;
   };
 
 bool IsThreePointsOnOneLine(const Point2D& p1, const Point2D& p2, const Point2D& p3);
