+++ include/vector.h
@@ -13,9 +13,9 @@ class Vector3D final
 public:
     Vector3D() = default;
     Vector3D(const Point3D& p0, const Point3D& p1);
-    Vector3D(float x, float y, float z) : x(x), y(y), z(z) {}
+    Vector3D(double x, double y, double z) : x(x), y(y), z(z) {}
 
-    float length() const;
+    double length() const;
     void normalize();
 
     bool valid() const;
@@ -25,19 +25,19 @@ public:
 
     Vector3D operator+(const Vector3D& vector) const;
     Vector3D operator-(const Vector3D& vector) const;
-    Vector3D operator*(float const_var) const;
+    Vector3D operator*(double const_var) const;
     void   operator=(const Point3D& point);
 
-    float x = NAN;
-    float y = NAN;
-    float z = NAN;
+    double x = NAN;
+    double y = NAN;
+    double z = NAN;
 };
 
 Vector3D Cross(const Vector3D& v1, const Vector3D& v2);
 
-float Dot(const Vector3D& v1, const Vector3D& v2);
+double Dot(const Vector3D& v1, const Vector3D& v2);
 
-float Dot(const Vector3D& v1, const Point3D& p2);
+double Dot(const Vector3D& v1, const Point3D& p2);
 
 //
 // Vector2D Block
@@ -48,9 +48,9 @@ class Vector2D final
 public:
     Vector2D() = default;
     Vector2D(const Point2D& p0, const Point2D& p1);
-    Vector2D(float x, float y) : x(x), y(y) {}
+    Vector2D(double x, double y) : x(x), y(y) {}
 
-    float length() const;
+    double length() const;
 
     bool valid() const;
     bool equal(const Vector2D& another_vector) const;
@@ -59,14 +59,14 @@ public:
 
     Vector2D operator+(const Vector2D& vector) const;
     Vector2D operator-(const Vector2D& vector) const;
-    Vector2D operator*(float const_var) const;
+    Vector2D operator*(double const_var) const;
     void   operator=(const Point2D& point);
 
-    float x = NAN;
-    float y = NAN;
+    double x = NAN;
+    double y = NAN;
 };
 
-float Cross(const Vector2D& v1, const Vector2D& v2);
+double Cross(const Vector2D& v1, const Vector2D& v2);
 
-float Dot(const Vector2D& v1, const Vector2D& v2);
+double Dot(const Vector2D& v1, const Vector2D& v2);
